Add --visible-pdf-image argument
This adds an optional command-line argument, `--visible-pdf-image`,
which is used as the image written to PDF when the the `pdf`
configfile is specified.

The normal input argument will be used for OCR processing. This can be
an image which has had filters applied to remove visual artifacts and
optimize for OCR efficiency. The "visible" image will be the one the
user sees when they open the PDF.

This patch applies on top of Tesseract 5.0.1.

Signed-off-by: David Hendricks <david.hendricks@gmail.com>